SP20: UART print time must be taken into account
When running in tsim/sis the UART output takes zero-time which makes
the test run successfully. On real hardware however the UART is
rather slow and it affects the execution time.

There is a switch to TSIM which makes the output printed to screen at
the same rate as the configured UART baudrate.

Signed-off-by: Daniel Hellstrom <daniel@gaisler.com>