b1553rt: Workaround for reset mode command issue
The B1553RT wrapper has been designed so that when a reset remote terminal
mode code is recieved, the whole wrapper is reset including the RT address
and IRQ mask.

Fortunately, the IRQ is delivered before this happens, so we can detect
this in the driver and restore the control register contents.