DRVMGR: reworked the address translation functions
Cleaned up the address translation implementation and updated
all users to the new interface. The old implementation had a bug
and the users used it incorrect. This patch aims to fix both
problems.

One can not translate a address upstreams and then take that
address and expect it to be translateable downstreams..