GR-RASTA-TMTC: updated for new version
From this driver's point of view the major new thing is that the
GRPCI peripheral PCI bridge has been updated to GRPCI2, the second
version. This means that both Big and Little Endian systems are now
supported and autodetected on runtime.

The PCI frequency is used as AMBA frequency of the GR-RASTA-TMTC.